Move vite plugins into src and share them across extension configs
