+++ extensions/jupyterlab/vite.config.ts
@@ -2,7 +2,7 @@ import { svelte } from '@sveltejs/vite-plugin-svelte'
 import { resolve } from 'node:path'
 import type { Plugin, PluginOption } from 'vite'
 import { defineConfig } from 'vite'
-import { three_compat_alias, vite_plugin_json_gz } from '../../vite-plugins.ts'
+import { three_compat_alias, vite_plugin_json_gz } from '../../src/vite-plugins.ts'
 
 const repo_root = resolve(import.meta.dirname, `../..`)
 
@@ -49,6 +49,9 @@ export default defineConfig({
     stub_vite_preload(),
     svelte({ compilerOptions: { runes: true } }),
   ] as PluginOption[],
+  // Worker bundles are a separate rolldown pass that does not inherit `plugins`, so the
+  // .json.gz loader must be registered again — structure-id's worker imports element data.
+  worker: { plugins: () => [vite_plugin_json_gz()] as PluginOption[] },
   build: {
     outDir: `lib`,
     target: `es2023`,
