+++ extensions/vscode/readme.md
@@ -144,13 +144,13 @@ MatterViz provides extensive customization options through VSCode settings. Acce
 
 ### Setting Categories
 
-| Category        | Description                     | Example Settings                                                                  |
-| --------------- | ------------------------------- | --------------------------------------------------------------------------------- |
-| **General**     | Global appearance and behavior  | `color_scheme`, `background_color`                                                |
-| **Structure**   | 3D structure visualization      | `atom_radius`, `bond_thickness`, `show_cell`, `ambient_light`, `show_image_atoms` |
-| **Trajectory**  | Animation and playback controls | `fps`, `auto_play`, `display_mode`, `show_controls`                               |
-| **Plots**       | Scatter plots and histograms    | `scatter.point.size`, `plot.grid_lines`, `plot.auto_fit_range`                    |
-| **Performance** | Memory and processing options   | `trajectory.chunk_size`, `trajectory.use_indexing`, `structure.sphere_segments`   |
+| Category        | Description                     | Example Settings                                                                          |
+| --------------- | ------------------------------- | ----------------------------------------------------------------------------------------- |
+| **General**     | Global appearance and behavior  | `color_scheme`, `background_color`                                                        |
+| **Structure**   | 3D structure visualization      | `atom_radius`, `bond_thickness`, `show_cell_vectors`, `ambient_light`, `show_image_atoms` |
+| **Trajectory**  | Animation and playback controls | `fps`, `auto_play`, `display_mode`, `show_controls`                                       |
+| **Plots**       | Scatter plots and histograms    | `scatter.point.size`, `plot.grid_lines`, `plot.auto_fit_range`                            |
+| **Performance** | Memory and processing options   | `trajectory.chunk_size`, `trajectory.use_indexing`, `structure.sphere_segments`           |
 
 ### Pro Tips
